refactor(xmlrpc): extract parsing tests to sibling file
